he movido todos los modelos a la carpeta src junto con la carpeta db y el archivo index.js y actualicé el código donde se hacen la importación para evitar errores - realicé el modelo de inscripción
