+++ package.json
@@ -1,15 +1,26 @@
 {
-  "name": "Orichalcum-proyecto-back",
+  "name": "orichalcum_back",
   "version": "1.0.0",
   "main": "index.js",
-  "repository": "https://github.com/kevcifu/Orichalcum-proyecto-back.git",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kevcifu/Orichalcum-proyecto-back.git"
+  },
   "author": "kevcifu <kevin.cifuentes@medellin.edu.co>",
   "license": "MIT",
   "scripts": {
-    "start": "node index.js"
+    "start": "nodemon src/index.js"
   },
   "dependencies": {
+    "apollo-server-express": "^3.5.0",
     "express": "^4.17.1",
-    "mongoose": "^6.0.14"
-  }
+    "mongoose": "^6.0.14",
+    "nodemon": "^2.0.15"
+  },
+  "bugs": {
+    "url": "https://github.com/kevcifu/Orichalcum-proyecto-back/issues"
+  },
+  "homepage": "https://github.com/kevcifu/Orichalcum-proyecto-back#readme",
+  "keywords": [],
+  "description": ""
 }
