+++ src/models/usuario.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require('mongoose');
-
 const userSchema = new Schema({
     correo: {
         type: String,
